LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-remove-expected.txt


FAIL First meta applies. assert_equals: expected "dark" but got "normal"
FAIL Second meta applies after first one is removed. assert_equals: expected "light" but got "normal"
PASS Initial color-scheme with both meta elements removed.
